Add required parameter annotations to eventList
Fixes #997

+++ spring-cloud-consul-binder/src/main/java/org/springframework/cloud/consul/binder/EventService.java
@@ -98,7 +98,7 @@ public class EventService {
 		if (this.properties != null) {
 			eventTimeout = this.properties.getEventTimeout();
 		}
-		ResponseEntity<List<Event>> watch = this.consul.eventList(eventTimeout, index);
+		ResponseEntity<List<Event>> watch = this.consul.eventList((long) eventTimeout, index);
 		return filterEvents(readEvents(watch), lastIndex);
 	}
